% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibR.R
\name{calibR}
\alias{calibR}
\title{Calibration by using raking ratio}
\usage{
calibR(Xs, d, total, q = rep(1, length(d)), max_iter = 500)
}
\arguments{
\item{Xs}{}

\item{d}{}

\item{total}{}

\item{q}{}

\item{max_iter}{}
}
\value{
Return the g-weights.
}
\description{
Calibration by using raking ratio
}
