% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calibRaking}
\alias{calibRaking}
\title{}
\usage{
calibRaking(Xs, d, total, q, max_iter = 500L, tol = 1e-09)
}
\arguments{
\item{X}{}

\item{k}{}

\item{tore}{}

\item{toreBound}{}
}
\value{

}
\description{

}
\details{

}
\seealso{

}
\author{
Raphaël Jauslin \email{raphael.jauslin@unine.ch}
}
