\name{OptimSampling-package}
\alias{OptimSampling-package}
\alias{OptimSampling}
\docType{package}
\title{
\packageTitle{OptimSampling}
}
\description{
\packageDescription{OptimSampling}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OptimSampling}
\packageIndices{OptimSampling}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OptimSampling}

Maintainer: \packageMaintainer{OptimSampling}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
