% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{statMatch_random}
\alias{statMatch_random}
\title{Title}
\usage{
statMatch_random(object, EPS = 1e-08, Z2)
}
\arguments{
\item{object}{output from the function \code{linkage}}
}
\value{

}
\description{
Title
}
\examples{
rm(list = ls())
N=10000
p=5
X=array(rnorm(N*p),c(N,p))
EPS= 1e-9

n1=1000
n2=1000
s1=srswor(n1,N)
s2=srswor(n2,N)


id1=(1:N)[s1==1]
id2=(1:N)[s2==1]

d1=rep(N/n1,n1)
d2=rep(N/n2,n2)

X1=X[s1==1,]
X2=X[s2==1,]

re=harmonize(X1,d1,id1,X2,d2,id2)
w1=re$w1
w2=re$w2
object = linkage(X1,id1,X2,id2,w1,w2)
out <- statMatch_random(object)

round(colSums(object$conc$weight*object$conc),3)
round(colSums(cbind(w1*X1,w2*X2)),3)
round(colSums((out$conc$weight/out$q)*out$conc),3)

}
